refactor: reorganize config, project, and auth file responsibilities
- Merge src/config.ts (prismic.config.json CRUD) into src/project.ts
  since both are project-level concerns
- Recreate src/config.ts as the CLI config directory module
  (CREDENTIALS_PATH, UPDATE_NOTIFIER_STATE_PATH)
- src/auth.ts now imports paths from src/config.ts instead of owning them
- Update all import paths across consumers
