+++ src/adapters/nuxt.ts
@@ -7,7 +7,7 @@ import { relative } from "node:path";
 import { fileURLToPath } from "node:url";
 
 import { Adapter } from ".";
-import { buildRoutePath, readConfig, updateConfig } from "../config";
+import { buildRoutePath, readConfig, updateConfig } from "../project";
 import { exists, writeFileRecursive } from "../lib/file";
 import { addDependencies, getNpmPackageVersion } from "../lib/packageJson";
 import { dedent } from "../lib/string";
