+++ src/index.ts
@@ -4,7 +4,8 @@ import { parseArgs } from "node:util";
 
 import packageJson from "../package.json" with { type: "json" };
 import { getAdapter, NoSupportedFrameworkError } from "./adapters";
-import { UPDATE_NOTIFIER_STATE_PATH, getHost, refreshToken } from "./auth";
+import { getHost, refreshToken } from "./auth";
+import { UPDATE_NOTIFIER_STATE_PATH } from "./config";
 import { getProfile } from "./clients/user";
 import docs from "./commands/docs";
 import gen from "./commands/gen";
@@ -18,7 +19,7 @@ import sync from "./commands/sync";
 import token from "./commands/token";
 import webhook from "./commands/webhook";
 import whoami from "./commands/whoami";
-import { InvalidPrismicConfig, MissingPrismicConfig } from "./config";
+import { InvalidPrismicConfig, MissingPrismicConfig } from "./project";
 import { CommandError, createCommandRouter } from "./lib/command";
 import {
 	ForbiddenRequestError,
